requirements.txt
opencv-python
sentence-transformers
faiss-cpu
numpy
qrcode[pil]
Pillow
tqdm
loguru
pyzbar
gradio 